+++ tasks.py
@@ -5,7 +5,6 @@ For now... might migrate to click.
 import datetime as dt
 import json
 from hashlib import sha256
-from pathlib import Path
 
 from invoke import task
 from rich.console import Console